Work on hooking Redux-Form to Redux-WebExt
We really don't want to write our own form library for WebExtensions.
It is proving challenging to integrate the Redux-WebExt library with
Redux-Forms because the actions from the Redux-Forms reducer are not
recognized by the "background store" logic in Redux-WebExt. I'm
not sure why you have to explicitly redefine the actions for the
background store.

Refs: FBBL-WE-5