Refit for Redux-Thunk
This set of commits fixes some naming conventions and
introduces redux-thunk for asynchronous processing. It
still has some processing going on that I know is not
being handled correctly but these commits begin to get
the basic classifier logic functionality visible into
the React/Redux interfaces, proving that it can all work.

I struggled for a bit with how to pass thunks across the
WebExtension proxy before realizing that this is exactly
what the action mapping is for. I will now proceed with
merging this branch because it does make a significant
improvement in the UI tools available.

Refs: FBBL-WE-5