Add corpura panel. Experiment with swipeview issue
The SwipableView panel seems to have problems with either
Redux connect or Redux-Form, I'm not quite sure which. Strangely,
if the first view in the swipeable collection is straight HTML
then everything (including connected redux-forms) works fine. As
much as I would like to run the problem all the way to ground I
am more interested in suppressing it so I can move forward.

Refs: FBBL-WE-17