Fixed menu layout, added db version
This is a continued refactoring of the UI to try to
deal with a "local storage first" type of user who does
not have access to a server to store their classifications.

This means that instead of syncing recipes and corpora
continuously with the server that we push and pull the data
on request. So you could pull a recipe from one server and
then push it to another. Currently the corpus and recipes
are keyed only on their name because it wasn't clear to me
how we should handle the problem beyond carefully chosen
names, perhaps something like Java class names.

Refs: FBBL-WE-21