New icon, refactored analyzer pipeline
The web extension now has a new icon and analysis basically
works. The analyzers are trained from their corpora on startup.
One current kink is that you have to reset the copora to get the
recipe to retrain if you alter the corpora. This will get fixed
in the future.